+++ src/main/java/com/verdantartifice/thaumicwonders/common/items/tools/ItemPrimalDestroyer.java
@@ -9,22 +9,30 @@ import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.init.MobEffects;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ItemSword;
+import net.minecraft.nbt.NBTTagInt;
 import net.minecraft.potion.PotionEffect;
+import net.minecraft.util.DamageSource;
+import net.minecraft.util.text.TextComponentString;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.EnumHelper;
 import net.minecraftforge.fml.common.FMLCommonHandler;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+import thaumcraft.api.capabilities.IPlayerKnowledge;
+import thaumcraft.api.capabilities.ThaumcraftCapabilities;
 import thaumcraft.api.items.IWarpingGear;
 import thaumcraft.api.items.ItemsTC;
 
 public class ItemPrimalDestroyer extends ItemSword implements IWarpingGear {
-    public static Item.ToolMaterial toolMatVoidflame = EnumHelper.addToolMaterial("VOIDFLAME", 4, 200, 8.0F, 6.0F, 20).setRepairItem(new ItemStack(ItemsTC.ingots, 1, 1));
+    public static final int MAX_HUNGER = 1800;
+    
+    public static Item.ToolMaterial toolMatVoidflame = EnumHelper.addToolMaterial("VOIDFLAME", 4, 200, 8.0F, 8.0F, 20).setRepairItem(new ItemStack(ItemsTC.ingots, 1, 1));
     
     public ItemPrimalDestroyer() {
         super(toolMatVoidflame);
@@ -41,8 +49,50 @@ public class ItemPrimalDestroyer extends ItemSword implements IWarpingGear {
     @Override
     public void onUpdate(ItemStack stack, World worldIn, Entity entityIn, int itemSlot, boolean isSelected) {
         super.onUpdate(stack, worldIn, entityIn, itemSlot, isSelected);
-        if (stack.isItemDamaged() && entityIn != null && (entityIn.ticksExisted % 20 == 0) && entityIn instanceof EntityLivingBase) {
-            stack.damageItem(-1, (EntityLivingBase)entityIn);
+        if (!worldIn.isRemote && entityIn != null && (entityIn.ticksExisted % 20 == 0) && entityIn instanceof EntityPlayer) {
+            EntityPlayer entityPlayer = (EntityPlayer)entityIn;
+            if (stack.isItemDamaged()) {
+                stack.damageItem(-1, entityPlayer);
+            }
+            
+            boolean inOffHand = itemSlot == 0 && ItemStack.areItemStacksEqual(stack, entityPlayer.getHeldItemOffhand());
+            boolean held = isSelected || inOffHand;
+            boolean onHotbar = (itemSlot >= 0 && itemSlot <= 8);
+            boolean equipped = held || onHotbar;
+            if (equipped) {
+                int hunger = 0;
+                if (stack.hasTagCompound()) {
+                    hunger = stack.getTagCompound().getInteger("hunger");
+                }
+                if (hunger >= MAX_HUNGER) {
+                    // Damage player and reset hunger
+                    entityPlayer.sendStatusMessage(new TextComponentString(TextFormatting.DARK_PURPLE + I18n.format("event.primal_destroyer.hunger_full")), true);
+                    entityPlayer.attackEntityFrom(new DamageSource("primalDestroyerHunger"), 12.0F);
+                    entityPlayer.addPotionEffect(new PotionEffect(MobEffects.WEAKNESS, 60));
+                    entityPlayer.addPotionEffect(new PotionEffect(MobEffects.HUNGER, 120));
+                    hunger = 0;
+                    
+                    // Give addenda research
+                    IPlayerKnowledge knowledge = ThaumcraftCapabilities.getKnowledge(entityPlayer);
+                    if (!knowledge.isResearchKnown("f_thevoidhungers")) {
+                        knowledge.addResearch("f_thevoidhungers");
+                        knowledge.sync((EntityPlayerMP)entityPlayer);
+                    }
+                } else {
+                    hunger++;
+                }
+                stack.setTagInfo("hunger", new NBTTagInt(hunger));
+            }
+        }
+    }
+    
+    @Override
+    public boolean shouldCauseReequipAnimation(ItemStack oldStack, ItemStack newStack, boolean slotChanged) {
+        if (oldStack.getItem() == newStack.getItem() && !slotChanged) {
+            // Suppress the re-equip animation if only the NBT data has changed
+            return false;
+        } else {
+            return super.shouldCauseReequipAnimation(oldStack, newStack, slotChanged);
         }
     }
 
@@ -56,7 +106,23 @@ public class ItemPrimalDestroyer extends ItemSword implements IWarpingGear {
                 target.setFire(3);
             } catch (Exception e) {}
         }
+        if (!target.world.isRemote) {
+            if (target.getHealth() <= 0.0F) {
+                this.decreaseHunger(stack, MAX_HUNGER);
+            } else {
+                this.decreaseHunger(stack, MAX_HUNGER / 5);
+            }
+        }
         return super.hitEntity(stack, target, attacker);
+    }
+    
+    private void decreaseHunger(ItemStack stack, int delta) {
+        int hunger = 0;
+        if (stack.hasTagCompound()) {
+            hunger = stack.getTagCompound().getInteger("hunger");
+        }
+        hunger = Math.max(0, hunger - delta);
+        stack.setTagInfo("hunger", new NBTTagInt(hunger));
     }
 
     @Override